Add a newScratch template func
Fixes #4685
